requirements.txt
poetry
fastapi
uvicorn
celery
redis
torch
torchaudio
librosa
numpy
pydantic
requests
python-dotenv
tenacity